One more 2018 Fellows card format
For future reference, I think that we should add collapsible content to these; they only include the first paragraph of bio info from the site, and we need to add the rest at some point. From what I've found, that's a JavaScript thing? So just something to keep in mind for later.